add: customer office rating endpoint

+++ DriveRent/cars/models.py
@@ -56,7 +56,7 @@ class Office(models.Model):
     location = models.CharField(max_length=255, verbose_name="موقع المكتب")
     image = models.ImageField(upload_to=PathAndRename('image_office'), blank=True, null=True)
     status_office=models.BooleanField(default=True)
-    ratings =models.PositiveSmallIntegerField(blank=True, null=True)
+    ratings =models.FloatField(blank=True, null=True)  
     phone_number_1 = models.CharField(max_length=20, verbose_name="رقم التواصل 1")
     phone_number_2 = models.CharField(max_length=20, verbose_name="رقم التواصل 2", blank=True, null=True)
 
@@ -86,7 +86,7 @@ class OfficeRating(models.Model):
     id_office_rating=models.AutoField(primary_key=True)
     customer = models.ForeignKey(Customer, on_delete=models.CASCADE)
     office = models.ForeignKey('Office', on_delete=models.CASCADE, verbose_name="المكتب")
-    rating = models.PositiveSmallIntegerField(blank=True, null=True)  
+    rating = models.FloatField(blank=True, null=True)  
     comment = models.TextField(blank=True, null=True)
     created_at = models.DateTimeField(auto_now_add=True)
